Improve migration

+++ shared/src/commonMain/kotlin/com/faltenreich/diaguard/preference/version/VersionCodePreference.kt
@@ -4,13 +4,13 @@ import com.faltenreich.diaguard.preference.Preference
 import diaguard.shared.generated.resources.Res
 import diaguard.shared.generated.resources.preference_version_code
 
-data object VersionCodePreference : Preference<Int, Int> {
+data object VersionCodePreference : Preference<Long, Long> {
 
     override val key = Res.string.preference_version_code
 
-    override val default = 0
+    override val default = 0L
 
-    override val onRead = { versionCode: Int -> versionCode }
+    override val onRead = { versionCode: Long -> versionCode }
 
-    override val onWrite = { versionCode: Int -> versionCode }
+    override val onWrite = { versionCode: Long -> versionCode }
 }
